refactor: 重构系统资源管理器相关接口
